Add Support for building Micropython using project generator (#704)
* Add Support for building Micropython using project generator (#702)

Micropython needs the C++ source files to have .cpp file extensions
where as Tensorflow uses the .cc extension.

Adds a new flag, --rename-cc-to-cpp,  to the project generator script
that will cause the copied .cc files to become .cpp files in the
destination directory.

Add a new test to the project generation tests.

* small updates to conform to the Google style for Python and shell.

* fix subtle issue with quoting the EXAMPLES string

* more fixes.

Co-authored-by: Advait Jain <advaitjain@google.com>
Co-authored-by: Advait Jain <advaitjain@users.noreply.github.com>