Port operator BROADCAST_ARGS from lite to micro  (#967)
* copy operator BROADCAST_ARGS kernel from lite

This is a copy without modification of the kernel and test for
operator FILL from tensorflow/lite/kernels.
Adaptations to micro and addition to the micro build to follow.

This follows the step in
main/tensorflow/lite/micro/docs/porting_reference_ops.md#5-port-the-op-from-lite-to-micro-pr3

* remove lite-specific code from copy of BROADCAST_ARGS

Remove the bulk of lite-specific code from the micro
implementation of operator BROADCAST_ARGS.
- Flatten the namespace
- Don't resize output tensors
- Don't use gtest

* Implement BROADCAST_ARGS operator for micro and add tests

* change 2021 to 2022